add platform info to JSON
